@@ -2,6 +2,7 @@ package com.highestaim.coronavirusinfo
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import com.highestaim.coronavirusinfo.databinding.ActivityMainBinding
 import com.highestaim.coronavirusinfo.fragment.HomeFragment
 import com.highestaim.coronavirusinfo.fragment.StatisticFragment
 
@@ -9,7 +10,7 @@ class MainActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
+        setContentView(ActivityMainBinding.inflate(layoutInflater).root)
 
         openMainPage()
     }
